+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
-## Current
+## Release 0.4.6
 - Fixed lost decoding memory at disconnection.
+- Fixed issue in send methods where sometimes data is lost.
+- Fixed lost fragmented UDP.
+  Before this change, a huge UDP message could be sent without respect the MTU size, now it panics.
+- Fixed rare issue of EventQueue dropping when the sender send and event and the receiver
+  is already removed.
 
 ## Release 0.4.5
 - Fixed enconding issue related several messages in the same data chunk.